backend/backend/requirements.txt
# ============================================
# SmartConvoy AI (SARATHI) - Backend Requirements
# Military Convoy Route Optimization System
# ============================================

# Web Framework & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1
python-multipart==0.0.6

# Database & PostgreSQL
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# HTTP Requests & APIs
requests==2.31.0

# Data Validation & Modeling
pydantic==2.5.3
pydantic[email]==2.5.3
typing-extensions==4.9.0

# Data Processing & Analysis
pandas==2.2.0
numpy==1.26.3

# Machine Learning (ETA Prediction)
scikit-learn==1.4.0
joblib==1.4.0

# Geospatial & Routing
polyline==2.0.0

# Mapping & Visualization
folium==0.15.1

# Environment Configuration
python-dotenv==1.0.0

# AI & Chatbot
google-generativeai==0.3.2
